requirements.txt
numpy
scipy
matplotlib
scikit-learn
pyyaml